Improved scons; Created pb_colloid, pb_double_plane and p_spherical_capacitor, with separate input files; code cleanup

+++ types.hpp
@@ -34,7 +34,7 @@ along with eke.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef TYPES_HPP
 #define TYPES_HPP
 
-typedef double Real;
+typedef long double Real;
 
 
 //// BLITZ ////
